core: add abort.{c,h} pager neutral processing
Moves structs, functions and defines which are central in abort
processing into abort.{c,h}. All identifier keeps their name in this
commit to make it easier to follow how the code is moved.
